+++ src/themes/teams/fontFaces.ts
@@ -1,0 +1,21 @@
+import { FontFaces } from '../../../types/theme'
+
+const fontFaces: FontFaces = [
+  {
+    name: 'Segoe UI',
+    paths: ['public/fonts/segoe-ui-regular.woff2'],
+    style: { fontWeight: 400 },
+  },
+  {
+    name: 'Segoe UI',
+    paths: ['public/fonts/segoe-ui-semibold.woff2'],
+    style: { fontWeight: 600 },
+  },
+  {
+    name: 'Segoe UI',
+    paths: ['public/fonts/segoe-ui-bold.woff2'],
+    style: { fontWeight: 700 },
+  },
+]
+
+export default fontFaces